Delete post actions on transitions when a workflow is deleted.

+++ plugins/tracker/include/workflow/WorkflowFactory.class.php
@@ -109,7 +109,27 @@ class WorkflowFactory {
      * @param int $workflow_id the workflow id
      */
     public function deleteWorkflowTransitions($workflow_id) {
-        return $this->getTransitionDao()->deleteWorkflowTransitions($workflow_id);
+        //The post actions on transitions must be deleted first
+        if ($this->deletePostActionsByTransition($workflow_id)) {
+            return $this->getTransitionDao()->deleteWorkflowTransitions($workflow_id);
+        }
+    }
+    
+    /**
+     * Delete the post actions of each transition of a workflow
+     *
+     * @param int $workflow_id the workflow id
+     */
+    public function deletePostActionsByTransition($workflow_id) {
+        $workflow = $this->getWorkflow($workflow_id);
+        $transitions = $this->getTransitions($workflow);
+        $tpaf = new Transition_PostActionFactory();
+        foreach ($transitions as $transition) {
+            if ($transition->getPostActions()) {
+                $tpaf->deletePostAction($transition);
+            }
+        }
+        return true;
     }
     
     /**
